% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flight_metadata.R
\name{add_flight_metadata}
\alias{add_flight_metadata}
\title{Add flight metadata}
\usage{
add_flight_metadata(flight_data, flight_id, flight_metadata)
}
\arguments{
\item{flight_data}{A processed flight data frame generated with the function `rnp_flight_data.R`}

\item{flight_id}{The ID of the flight to match its metadata.}

\item{flight_metadata}{A flight metadata data frame.}
}
\value{
A flight data frame with metadata.
}
\description{
Adds metadata (e.g., flight time) to a processed flight data frame
}
