% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_otm_metadata.R
\name{add_otm_metadata}
\alias{add_otm_metadata}
\title{Add OTM metadata}
\usage{
add_otm_metadata(otm_data, otm_metadata)
}
\arguments{
\item{otm_data}{An OTM data frame generated using the `rnp_otm_data.R` function.}

\item{otm_metadata}{An OTM metadata data frame with an otm_id column to match with `otm_data`}
}
\value{
An OTM data frame with metadata included
}
\description{
Adds metadata (e.g., location, orientation, microhabitat characterstis) to an existing OTM dataset
}
