% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_correction.R
\name{apply_correction}
\alias{apply_correction}
\title{Apply correction}
\usage{
apply_correction(flight_data, correction_factors)
}
\arguments{
\item{flight_data}{A flight data frame with metadata that can be a collection of multiple flights binded together.}

\item{correction_factors}{A correction factor list obtain used the `get_correction_factors` function.}
}
\value{
The same flight data with an operative temperature column
}
\description{
Applies a correction to an existing flight data frame to transform it into operative temperature
}
