% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{bad_read_otm}
\alias{bad_read_otm}
\title{OTM raw data file read incorrectly}
\format{
An object of class \code{data.frame} with 6159 rows and 1 columns.
}
\usage{
bad_read_otm
}
\description{
Example used in the vignettes of the package to illustrate what happens when
a raw OTM file is read incorrectly.
}
\examples{
head(bad_read_otm)
}
\keyword{datasets}
