% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_flights_data.R
\name{correct_flights_data}
\alias{correct_flights_data}
\title{Correct flights data}
\usage{
correct_flights_data(flights_data, otm_splines)
}
\arguments{
\item{flights_data}{A `tibble` of flights data obtained through the `rnp_flights_data`
function. The `tibble` must contain columns for `longitude` and `latitude`.}

\item{otm_splines}{A complex `tibble` obtained using the `gen_otm_splines` function.
The `tibble` must contain columns for `longitude` and `latitude` and all values
must also be in the `flights_data`}
}
\value{
A processed `flights_data` `tibble` where IR-measured temperatures (`ir_temp`)
   has been corrected to operative temperatures (`op_temp`).
}
\description{
Correct flights data to transform IR-measured temperatures into operative temperatures
}
