% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{correction_data}
\alias{correction_data}
\title{A \code{tibble} with a comparison between the surface temperatures recorded
by flights on tiles where OTM were deployed and the operative temperatures
recorded by OTMs during the exact same period the flights were taking place}
\format{
A \code{tibble} of 1114 rows and 8 columns:
\describe{
\item{latitude}{A numeric column indicating latitude of a tile}
\item{longitude}{A numeric column indicating longitude of a tile}
\item{year}{A numeric columnindicating the year when the flight took place}
\item{doy}{A numeric column indicating the day of the year when the flight took place}
\item{mod_start}{A numeric column indicating the minute of the day when the flight started}
\item{mod_end}{A numeric column indicating the minute of the day when the flight ended}
\item{ir_temp}{A numeric column indicating the average surface temperature recorded in a tile in a given flight}
\item{op_temp}{A numeric column indicating the average operative temperature recorded by an OTMs in a tile}
}
}
\usage{
correction_data
}
\description{
A \code{tibble} with a comparison between the surface temperatures recorded
by flights on tiles where OTM were deployed and the operative temperatures
recorded by OTMs during the exact same period the flights were taking place
}
\examples{
correction_data
}
\keyword{datasets}
