% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{elevation}
\alias{elevation}
\title{Elevation profile of the area where we conducted the project at a spatial
resolution of ~ 1m^2}
\format{
A \code{tibble} with 8170 rows and 3 columns, all numeric:
\describe{
 \item{longitude}{Longitude in decimal degrees}
 \item{latitude}{Latitude in decimal degrees}
 \item{elevation}{Elevation in meters}
}
}
\usage{
elevation
}
\description{
Elevation profile of the area where we conducted the project at a spatial
resolution of ~ 1m^2
}
\examples{
elevation
}
\keyword{datasets}
