% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{flights_metadata}
\alias{flights_metadata}
\title{Metadata for example flights}
\format{
A data frame with 34 rows and 4 columns:
\describe{
 \item{flight_id}{Character column with a unique identifier for each flight}
 \item{date}{Character column with the date when the flight took place in DD/MM/YYYY format}
 \item{time_start}{Character column with the time when the flight started in HH:MM format}
 \item{time_end}{Character column with the time when the flight ended in HH:MM format}
 }
}
\usage{
flights_metadata
}
\description{
Metadata for example flights
}
\examples{
tibble::as_tibble(flights_metadata)
}
\keyword{datasets}
