% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_corr_data.R
\name{gen_corr_data}
\alias{gen_corr_data}
\title{Generate validation data}
\usage{
gen_corr_data(flight, splines)
}
\arguments{
\item{flight}{A flight data frame}

\item{splines}{An OTM splines object generated with `gen_otm_splines`}
}
\value{
A validation dataset
}
\description{
Generates a validation dataset
}
