% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_otm_splines.R
\name{gen_otm_splines}
\alias{gen_otm_splines}
\title{Generate OTM splines}
\usage{
gen_otm_splines(otm_data, knot_p)
}
\arguments{
\item{otm_data}{An OTM data \code{tibble} obtained via the function
\code{rnp_otms_data}.}

\item{knot_p}{The number of knots the spline model should have to describe the
temperature dynamics of the OTM for a given day taken as a percentage of
the total number of operative measurements in a given day.}
}
\value{
A nested data frame (\code{tibble}) with columns for \code{otm_id},
   \code{year}, \code{doy} and \code{spline}. The spline column is nested and
   contains a OTM, year and DOY \code{smooth.spline} model with the specified
   number of knots.
}
\description{
Generate a unique spline model for each OTM on each date (Julian) while the OTM was measuring
}
