% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_correction_factors.R
\name{get_correction_factors}
\alias{get_correction_factors}
\title{Get correction estimates}
\usage{
get_correction_factors(corr_data)
}
\arguments{
\item{corr_data}{A correlation dataset obtained using the `gen_corr_data` function.}
}
\value{
A list of three objects, first being the correction by date & time dataset, second an intercept and third a slope for the general temperature correction
}
\description{
Gets correction estimates from for a correlation dataset
}
