% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utm_zone.R
\name{get_utm_zone}
\alias{get_utm_zone}
\title{Get UTM zone from a longitude value}
\usage{
get_utm_zone(longitude)
}
\arguments{
\item{longitude}{A decimal longitude value between 180 and -180. We recommend this being the average longitude of your site.}
}
\value{
An integer between 1 and 60 indicating the UTM zone
}
\description{
Returns the universal transverse mercator (UTM) projection zone of a given longitudinal value
}
\details{
Determined using the equation:
\deqn{UTM_{zone} = \lfloor \frac{longitude + 180}{6} \rfloor + 1}
}
\examples{

get_utm_zone(longitude = -119.45)
get_utm_zone(longitude = 2)

}
