% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{good_read_otm}
\alias{good_read_otm}
\title{OTM raw data file read correctly}
\format{
An object of class \code{data.frame} with 2048 rows and 3 columns.
}
\usage{
good_read_otm
}
\description{
Example used in the vignettes of the package to illustrate what happens when
a raw OTM file is read correctly
}
\examples{
head(good_read_otm)
}
\keyword{datasets}
