% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_data.R
\name{match_data}
\alias{match_data}
\title{Match flights to OTM data}
\usage{
match_data(flights_data, otm_splines, coverage_per, error_max)
}
\arguments{
\item{flights_data}{A \code{tibble} of flights data obtained through \code{rnp_flights_data}
and corrected using \code{correct_flights_data}. The temperature
column should be \code{op_temp} but the function will also work if \code{ir_temp} is provided.}

\item{otm_splines}{A nested \code{tibble} obtained using \code{gen_otm_splines}}

\item{coverage_per}{A numeric between 0 - 1 indicating the minimum coverage that
a tile should have across all flights provided in order to be included in the matching.
Coverage is calculated as the number of times temperature was measured in a given tile
divided by the total number of flights. Values >= 0.9 are recommended.
The function will provide a warning of the number of tiles for which
coverage is > 0.5.}

\item{error_max}{The maximum average absolute error between temperature measurements
of a tile and an OTM that makes a match between a tile and OTM valid.
Error is calculated as the average absolute value between the OTM prediction
and the temperature measurements of the tile.}
}
\value{
A matches \code{tibble} with columns for \code{latitude}, \code{longitde}, the \code{otm_id}
   that best describes the thermal dynamics of that tile and the average absolute
   \code{error} between tile measurements and OTM predictions. Rows where \code{is.na(otm_id)}
   indicate tiles where \code{error} was not \code{< error_max} for any of the OTMs provided.
}
\description{
Matches thermal dynamics of specific tiles (i.e., specific latitude and longitudes)
collected across multiple flights to the thermal dynamics of an operative
temperature model (OTM).
}
