% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{matches}
\alias{matches}
\title{Example matches data}
\format{
A \code{data.frame} of 6236 rows and 5 columns
\describe{
\item{x}{A numeric column indicating the horizontal UTM coordinate}
\item{y}{A numeric column indicating the vertical UTM coordinate}
\item{otm_id}{A character column indicating the OTM that best describe the dynamics of a tile,
 a value of \code{NA} indicates that that tile could not be matched with any OTM with an
 \code{error < error_max}}
\item{error}{A numeric column indicating the average absolute error between the
thermal dynamics of a tile and the OTM that best described it}
\item{coverage}{A numeric column indicating the coverage of a tile across all flights}
}
}
\usage{
matches
}
\description{
A \code{data.frame} indicating the OTM that best described the dynamics of a
given tile. This data was generated using the \code{match_data} function using
a flights data set processed using \code{rnp_flights_data} with a spatial resolution
of 1 m2 (the \code{flights_data} object included in \code{throne}), an OTM splines
\code{tibble} generated using \code{gen_otm_splines} setting the
\code{knot_p = 0.02} built from the \code{otms_data} data set
that is also included in \code{throne}, and by setting \code{error_max = 100}
and \code{coverage_per = 0.9}.
}
\examples{
matches
}
\keyword{datasets}
