% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{matches_100}
\alias{matches_100}
\title{Matches data with \code{error_max = 100}}
\format{
A \code{tibble} of 6386 rows and 4 columns
\describe{
\item{latitude}{A numeric column indicating latitude of a tile}
\item{longitude}{A numeric column indicating longitude of a tile}
\item{otm_id}{A character column indicating the OTM that best describe the dynamics of a tile,
 a value of \code{NA} indicates that that tile could not be matched with any OTM with an
 \code{error < error_max}}
\item{error}{A numeric column indicating the average absolute error between the
thermal dynamics of a tile and the OTM that best described it}
}
}
\usage{
matches_100
}
\description{
A \code{tibble} indicating the OTM that best describe the dynamics of a given tile
assuming a \code{error_max = 100} and a \code{coverage_per = 0.9} (i.e., default settings).The increased \code{error_max}
is for visualization purposes (i.e., so no empty tiles appear in the plots presented
in the documentation vignettes).The \code{flights_data} used to perform
the matching was corrected using the \code{correct_flights_data} function
specifying that both a time and a flight-specific temperature correction
should be performed.
}
\examples{
matches_100
}
\keyword{datasets}
