% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{matches_20}
\alias{matches_20}
\title{A \code{tibble} indicating the OTM that best describe the dynamics of a given tile
assuming a \code{error_max = 20}. The purpose of the increased error is for visualization
purposes (i.e., so no empty tiles appear in the plots presented in the documentation
vignettes).}
\format{
A \code{tibble} of 6386 rows and 4 columns
#' \describe{
\item{latitude}{A numeric column indicating latitude of a tile}
\item{longitude}{A numeric column indicating longitude of a tile}
\item{otm_id}{A character column indicating the OTM that best describe the dynamics of a tile,
 a value of \code{NA} indicates that that tile could not be matched with any OTM with an
 \code{error < error_max}}
\item{error}{A numeric column indicating the average absolute error between the
thermal dynamics of a tile and the OTM that best described it}
}
}
\usage{
matches_20
}
\description{
A \code{tibble} indicating the OTM that best describe the dynamics of a given tile
assuming a \code{error_max = 20}. The purpose of the increased error is for visualization
purposes (i.e., so no empty tiles appear in the plots presented in the documentation
vignettes).
}
\examples{
matches_5
}
\keyword{datasets}
