% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{otms_data}
\alias{otms_data}
\title{Example OTMs data}
\format{
A \code{data.frame} with 67584 rows and 10 columns
\describe{
\item{otm_id}{Character column with a unique identifier for each OTM}
\item{year}{A numeric column indicating year}
\item{doy}{A numeric column indicating day of the year}
\item{mod}{A numeric column indicating minute of the day}
\item{op_temp}{A numeric column indicating operative temperature}
\item{microhabitat}{Character column describing the microhabitat where the OTM was deployed,
 between \code{"outrcrop"}, \code{"small_rock"}, \code{"rock"},
 \code{"boulder"}, \code{"ground"}, and \code{"bush"}}
\item{orientation}{Character column describing the orientation at which the OTM was deployed,
 between \code{"N"}, \code{"NE"}, \code{"E"}, \code{"SE"}, \code{"S"}, \code{"SW"},
 \code{"W"}, \code{"NW"} and \code{"Flat"}}
}
}
\usage{
otms_data
}
\description{
A data set of 36 OTMs deployed in the field recording for 3 days at a rate of
1 observation every 2 minutes. This data set was obtained using the \code{rnp_otms_data}
function and incorporates some metadata for each OTM. Each row corresponds to
a unique operative temperature measurement recorded by a unique OTM at a given
moment in time (i.e., unique year, \code{doy} and \code{mod}).
}
\examples{
as_tibble(otms_data)
}
\keyword{datasets}
