% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{otms_metadata}
\alias{otms_metadata}
\title{Metadata for example OTMs}
\format{
A \code{tibble} with 36 rows and 6 columns:
\describe{
 \item{otm_id}{Character column with a unique identifier for each OTM}
 \item{microhabitat}{Character column describing the microhabitat where the OTM was deployed,
 between \code{"outrcrop"}, \code{"small_rock"}, \code{"rock"},
 \code{"boulder"}, \code{"ground"}, and \code{"bush"}}
 \item{orientation}{Character column describing the orientation at which the OTM was deployed,
 between \code{"N"}, \code{"NE"}, \code{"E"}, \code{"SE"}, \code{"S"}, \code{"SW"},
 \code{"W"}, \code{"NW"} and \code{"Flat"}}
 \item{latitude}{Latitude where the OTM was deployed in decimal degrees}
 \item{longitude}{Longitude where the OTM was deployed in decimal degrees}
 \item{elevation}{Elevation where the OTM was deployed in meters}
 }
}
\usage{
otms_metadata
}
\description{
Metadata for example OTMs
}
\examples{
otms_metadata
}
\keyword{datasets}
