% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{otms_splines}
\alias{otms_splines}
\title{Example OTM splines nested data}
\format{
A \code{tibble} of 132 rows and 9 columns:
\describe{
\item{otm_id}{Character column with a unique identifier for each OTM}
\item{year}{A numeric column indicating year}
\item{doy}{A numeric column indicating day of the year}
\item{microhabitat}{Character column describing the microhabitat where the OTM was deployed,
 between \code{"outrcrop"}, \code{"small_rock"}, \code{"rock"},
 \code{"boulder"}, \code{"ground"}, and \code{"bush"}}
\item{orientation}{Character column describing the orientation at which the OTM was deployed,
 between \code{"N"}, \code{"NE"}, \code{"E"}, \code{"SE"}, \code{"S"}, \code{"SW"},
 \code{"W"}, \code{"NW"} and \code{"Flat"}}
\item{latitude}{Latitude where the OTM was deployed in decimal degrees}
\item{longitude}{Longitude where the OTM was deployed in decimal degrees}
\item{elevation}{Elevation where the OTM was deployed in meters}
\item{spline}{A nested column containing the smoothing spline model}
}
}
\usage{
otms_splines
}
\description{
A nested \code{tibble} with smoothing splines models for each OTM during each
day while deployed in the field to validate the methodology of this package.
Each row corresponds to a unique OTM and day of the year combination. The data
set contains metadata information on the deployment of the OTMs. This data set
was generated from the data set \code{otms_data} using the \code{gen_otm_splines}
function specifying \code{knot_p = 4}.
}
\examples{

otms_splines

# when unnesting a spline model
otms_splines$spline[[2]]
}
\keyword{datasets}
