% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_thermal_landscape.R
\name{predict_thermal_landscape}
\alias{predict_thermal_landscape}
\title{Predict thermal landscape}
\usage{
predict_thermal_landscape(matches, otm_splines, date, minute)
}
\arguments{
\item{matches}{x}

\item{otm_splines}{x}

\item{date}{x}

\item{minute}{x}
}
\value{
x
}
\description{
Predict thermal landscape
}
