% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_n_process_flight_data.R
\name{read_n_process_flight_data}
\alias{read_n_process_flight_data}
\title{Read and process .TIF files}
\usage{
read_n_process_flight_data(file_path, utm_zone, hemisphere, digits)
}
\arguments{
\item{file_path}{The location of the .tif file as a character.}

\item{utm_zone}{The universal transverse mercator (UTM) zone where the drone flight took place.}

\item{hemisphere}{The hemisphere where the drone flight took place.}

\item{digits}{The number of decimal digits to which coordinates should be rounded to. Will define the spatial accuracy of the flight}
}
\value{
A processed flight object: A tibble with latitude (lat), longitude (long) and temperature (temp) as columns. Lat and long will have as many digits as specified above and temp will be in degrees C.
}
\description{
Reads and processes a raw flight data file (.TIF) it into a processed flight tibble object
}
