% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_n_process_tif.R
\name{read_n_process_tif}
\alias{read_n_process_tif}
\title{Read and process .tif files}
\usage{
read_n_process_tif(file_path, utm_zone, hemisphere, digits)
}
\arguments{
\item{file_path}{The location of the .tif file as a character.}

\item{utm_zone}{The universal transverse mercator (UTM) zone where the drone flight took place.}

\item{hemisphere}{The hemisphere where the drone flight took place.}

\item{digits}{The number of decimal digits to which coordinates should be rounded to. Will define the spatial accuracy of the flight}
}
\value{
A tibble with latitude (lat), longitude (long) and temperature (temp) as columns. Lat and long will have as many digits as specified above and temp will be in degrees C.
}
\description{
Reads a .tif file obtained from a drone flight and processes it into a tibble object
}
