% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnp_flight_data.R
\name{rnp_flight_data}
\alias{rnp_flight_data}
\title{Read and process flight data}
\usage{
rnp_flight_data(file_path, digits)
}
\arguments{
\item{file_path}{File path to flight}

\item{digits}{Number of decimal digits to round coordinates to}
}
\value{
Pending
}
\description{
Read and process thermal imaging data from a drone flight (in `.tif` format)
}
