% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnp_otm_data.R
\name{rnp_otm_data}
\alias{rnp_otm_data}
\title{Read and process OTM data}
\usage{
rnp_otm_data(folder_path, rows_skip, time_col, op_temp_col)
}
\arguments{
\item{folder_path}{Folder where raw OTM files (in .csv format) are stored.}

\item{rows_skip}{Rows to skip on each OTM file to get to the data.}

\item{time_col}{Column within an OTM file where the "time" variable is stored.}

\item{op_temp_col}{Column within an OTM file where the operative temperature ("op_temp") variable is stored.}
}
\value{
A data frame with otm_id, year, date (Julian), minute (of the day) and operatie temperature.
}
\description{
Reads OTM files from a specified folder and returns a processed data frame.
}
