% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_flights_data.R
\name{correct_flights_data}
\alias{correct_flights_data}
\title{Correct flights data}
\usage{
correct_flights_data(flights_data, otm_splines)
}
\arguments{
\item{flights_data}{A \code{tibble} of flights data obtained through the \code{rnp_flights_data}
function. The \code{tibble} must contain columns for `longitude` and `latitude`.}

\item{otm_splines}{A nested \code{tibble} obtained using the \code{gen_otm_splines} function.
The \code{tibble} must contain columns for \code{longitude} and \code{latitude} and all values
must also be in the \code{flights_data}}
}
\value{
A processed \code{flights_data} \code{tibble} where IR-measured temperatures (\code{ir_temp})
   has been corrected to operative temperatures (\code{op_temp}).
}
\description{
Correct flights data to transform IR-measured temperatures into operative temperatures
}
