% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_data.R
\name{match_data}
\alias{match_data}
\title{Match data}
\usage{
match_data(flights_data, coverage, otm_splines)
}
\arguments{
\item{flights_data}{A flight data frame for multiple flights}

\item{coverage}{Number of times a given pixel needs to be covered by a flight for it to be considered.}

\item{otm_splines}{An OTM splines dataset obtained with the `gen_otm_splines` function}
}
\value{
A match dataset where each pixel is associated with an OTM.
}
\description{
Matches each pixel within a given are where flights have occured with an OTM.
}
