% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{data}
\name{matches_5}
\alias{matches_5}
\title{Matches data with \code{error_max = 5}}
\format{
A \code{tibble} of 6386 rows and 4 columns
\describe{
\item{latitude}{A numeric column indicating latitude of a tile}
\item{longitude}{A numeric column indicating longitude of a tile}
\item{coverage}{The percentage of flights in which that tile was coverged}
\item{error}{A numeric column indicating the average absolute error between the thermal dynamics of a tile and the OTM that best described it}
\item{otm_id}{A character column indicating the OTM that best describe the dynamics of a tile,
 a value of \code{NA} indicates that that tile could not be matched with any OTM with an
 \code{error < error_max}}
}
}
\usage{
matches_5
}
\description{
A \code{tibble} indicating the OTM that best describe the dynamics of a given tile
assuming a \code{error_max = 5} and a \code{coverage_per = 0.9} (i.e., default settings).
The \code{flights_data} used to perform the matching was corrected using t
he \code{correct_flights_data} function specifying that both a time and a flight-specific
temperature correction should be performed.
}
\examples{
matches_5
}
\keyword{datasets}
