% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_thermal_landscape.R
\name{predict_thermal_landscape}
\alias{predict_thermal_landscape}
\title{Predict thermal landscape}
\usage{
predict_thermal_landscape(matches, otm_splines, doy, mod)
}
\arguments{
\item{matches}{A matches data frame obtained using the \code{match_data} function.}

\item{otm_splines}{A nested \code{tibble} obtained using the \code{gen_otm_splines} function.}

\item{doy}{An integer or vector of integers indicating the day(s) of the year
(\code{doy}) for which the thermal landscape should be predicted. All elements within
\code{doy} must be within the range between when the OTMs started and finished
recording operative temperatures.}

\item{mod}{A numeric vector of length >=1 indicating the minute(s) of the
day (\code{mod}) for which the thermal landscape should be predicted. All elements
within \code{mod} must fall within the 0 to 1440 range. If \code{mod} includes
values when the OTMs were not recording the function will remove them and provide
a diagnostics message indicating the percentage of predictions removed.}
}
\value{
A data frame with a predicted thermal landscape with \code{x}, \code{y},
   (UTM coordinates), \code{doy},\code{mod} and predicted operative temperature
   columns (\code{pred_op_temp}).
}
\description{
Predicts a thermal landscape by integrating information from an OTM-tile match
object and an \code{otm_splines} object for a specified range of days of the
year (\code{doy}) and minutes of the day (\code{mod}).
}
