% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_thermal_landscape.R
\name{predict_thermal_landscape}
\alias{predict_thermal_landscape}
\title{Predict thermal landscape}
\usage{
predict_thermal_landscape(matches, otm_splines, doy, mod)
}
\arguments{
\item{matches}{A matches \code{tibble} obtained using the \code{match_data} function.}

\item{otm_splines}{A complex \code{tibble} obtained using the \code{gen_otm_splines} function.}

\item{doy}{An integer or vector of integers indicating the day(s) of the year
(\code{doy}) for which the thermal landscape should be predicted. All elements within
\code{doy} must be within the range between when the OTMs started and finished
recording operative temperatures.}

\item{mod}{An integer of vector of integers indicating the minute(s) of the
day (MOD) for which the thermal landscape should be predicted. All elements
within \code{mod} must fall within the 0 to 1440 range. If \code{mod} includes
values when the OTMs were not recording it will return \code{NA} and provide a
diagnostics message}
}
\value{
A \code{tibble} with a predicted thermal landscape with \code{longitude}, \code{latitude},
   \code{doy}, \code{mod} and predicted operative temperature columns (\code{pred_op_temp}).
}
\description{
Predict thermal landscape
}
