% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnp_flight_data.R
\name{rnp_flight_data}
\alias{rnp_flight_data}
\title{Read and process flight data}
\usage{
rnp_flight_data(file_path, digits)
}
\arguments{
\item{file_path}{The path to the folder where the `.tif` file is stored.}

\item{digits}{The number of decimal digits to which the final output should be summarized to.}
}
\value{
Pending
}
\description{
Read and process thermal imaging data from a drone flight (in `.tif` format)
}
