% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnp_flights_data.R
\name{rnp_flights_data}
\alias{rnp_flights_data}
\title{Read and process flights data}
\usage{
rnp_flights_data(path, metadata, digits)
}
\arguments{
\item{path}{A character indicating the directory where all `.tif`
files for flights are stored. The name of each flight `.tif` file must
have associated metadata for the function to work.}

\item{metadata}{A metadata `tibble` for the flights with information on
the `flight_id` (character), `date` (MM/DD/YYYY), `time_start`(HH:MM)
and `time_end`(HH:MM) of each flight.}

\item{digits}{An integer (>0 & <6), indicating the number of decimal digit places
to which the final output should be summarized to. Will dictate the spatial
resolution of each of the tiles in the eventual thermal landscape.}
}
\value{
A `tibble` where each row represents a unique tile (a unique combination
   of `longitude` and `latitude`), with an associated IR temperature measurement
   (`ir_temp`) for a unique flight. To identify each flight, each row of the
   return `tibble` also contains information on the `year`, date of the year (`doy`)
   and the minute of the da (MOD) when the flight started (`mod_start`) and ended (`mod_end`)
}
\description{
Reads and processes data from multiple flights in `.tif` format and adds metadata
}
